Add new Visual Studio solution file
A new Visual Studio solution file (`myMLapp.sln`) has been created and added to the project. This file is in format version 12.00 and is compatible with Visual Studio Version 17. It includes metadata about the Visual Studio version used (`17.12.35527.113 d17.12`) and the minimum required Visual Studio version (`10.0.40219.1`). Additionally, it contains a global section for solution properties, specifying that the solution node should not be hidden.